New module kas-core::event::cx::accessibility
